feat: [#60305] Add `HarborDuesFormReceipt` and `CruiseTaxFormReceipt`
This adds the HTML templates used for rendering the PDF receipts.
These templates use Bootstrap CSS as well as local print styles.

The base template for receipts is dynamic, i.e. it can be defined by the
code rendering the receipt. The intention is to support rendering the
same receipts as both HTML and PDF by switching the base template.

This commit adds model properties used for template rendering:
* `HarborDuesForm.duration_in_days`,
* `HarborDuesForm.duration_in_weeks`,
* `CruiseTaxForm.total_tax`