fix: [#60305] Improve cruise form layouts
This simplifies the layouts of steps 2 and 3 when reporting cruise
taxes.

Both pages consist of rather long lists, where the user can fill out a
"number of passengers" for one or more nationalities or disembarkment
sites.

We simplify the layout by leaving out redundant labels and using the
nationality or disembarkment site as the label for each input instead.

This also removes the confusing "Cancel" buttons from each form step (1,
2 and 3.)