chore: [#60278] Update `shipping_agent_and_forms.json` fixture
The fixture now creates a separate unused `ShippingAgent` ("En anden
agent") so that the following ownerships can be established:

1) `HarborDuesForm` "Gøglerbåden" can be accessed by `User`
"portauthority", as its port of call is linked to the same port
authority ("Royal Arctic Line".)

2) `HarborDuesForm` "Gøglerbåden" cannot be accessed by `User`
"shipping" as it is linked to a different shipping agent ("En anden
agent".)

3) `HarborDuesForm` "Mary" can be accessed by `User` "shipping" as it is
linked to the same shipping agent ("Agenten".)

4) `CruiseTaxForm` "SkibSkib" is set up opposite to "Gøglerbåden":
accessible by `User` "portauthority" as they share the same port
authority, *and* by `User` "shipping" as they share the same shipping
agent.

5) `CruiseTaxForm` "Skibsnavn" is not linked to any port authority (it
is linked to a port without a port authority) and thus is not accessible
to `User` "portauthority". It is linked to `ShippingAgent` "Agenten" and
thus accessible to `User` "shipping".