feat: [#61559] Add "password reset" flow (URLs, templates, whitelisting)
The templates use the default English texts from the default Django
"password reset" templates, so that we can rely on existing Danish
translations. This means that we do not have to fill out the
corresponding entries in the `da` locale.

However, there are no Greenlandic translations in Django for these
strings. Currently, users who select Greenlandic will see English texts
for the password reset flow.