feat: [#61401] Add "approve" and "reject" modals/forms/views
The views only receive a POST payload from the modals - they do not
render a template on their own.

The views validate that a status transition is allowed by way of
`has_transition_perm` from `django-fsm`.

Only "submitters" (ship users and shipping agent users) are allowed to
submit forms for review (== setting status to `NEW`.)

Only port authority users are allowed to approve or reject forms (==
setting status to `APPROVED` or `REJECTED`, respectively.)

The modals are implemented as Django inclusion tags, so they can be
reused in other parts of the UI, if necessary.

The template context for a receipt now include the boolean variables
`can_approve` and `can_reject`, which determine whether the "approve"
and "reject" actions are available in the UI.