fix: [#61919] Render rejection messages for cruise tax forms as well
`RejectView` currently always saves the rejection reason via
`HarborDuesForm`, even if the object being rejected is a
`HarborDuesForm`. Thus, the rejection message is saved in
`HarborDuesForm.history` and not `CruiseTaxForm.history`.

This fix enables `HarborDuesForm.latest_rejection` to return the
rejection message stored in `HarborDuesForm.history` even if the object
in question is a `CruiseTaxForm`.