feat: [#64450] Remove `vessel type` from `UpdateVesselView`
Editing the vessel type now implies a new check of the IMO/nickname,
depending on the new vessel type.

Since the IMO/nickname can not be edited on `UpdateVesselView`, and thus
be made valid according to the rule for the new vessel type chosen, it
does not make sense to edit on the vessel type on its own.

If a ship user *must* revise their vessel type, IMO number/nickname, and
username, they must reach out to us, and we can do it via the Django
admin.